Chore: 문의하기 코드 수정

+++ src/main/java/com/palleteforco/palleteforco/domain/inquiry/service/InquiryServiceImpl.java
@@ -2,9 +2,12 @@ package com.palleteforco.palleteforco.domain.inquiry.service;
 
 import com.palleteforco.palleteforco.domain.inquiry.dto.InquiryDto;
 import com.palleteforco.palleteforco.domain.inquiry.mapper.InquiryMapper;
+import com.palleteforco.palleteforco.global.exception.ForbiddenException;
 import com.palleteforco.palleteforco.global.exception.NotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,7 +27,13 @@ public class InquiryServiceImpl implements InquiryService {
     // 문의사항 등록
     @Transactional
     public void registerInquiry(InquiryDto inquiryDto) throws Exception {
-        inquiryDto.setInquiry_regdate(LocalDateTime.now());
+        OAuth2User oAuth2User = (OAuth2User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String email = (String)oAuth2User.getAttributes().get("email");
+
+        inquiryDto.setEmail(email);
+
+        inquiryDto.setInquiry_reg_date(LocalDateTime.now());
+
         inquiryMapper.insertInquiry(inquiryDto);
     }
 
@@ -39,7 +48,7 @@ public class InquiryServiceImpl implements InquiryService {
         InquiryDto inquiryDto = inquiryMapper.selectInquiryListDetail(inquiry_id);
 
         if (inquiryDto == null) {
-            throw new NotFoundException("찾으시는 게시물이 없습니다.");
+            throw new NotFoundException("해당 게시물이 없습니다.");
         }
 
         inquiryDto.setInquiry_update_date(LocalDateTime.now());
@@ -56,6 +65,13 @@ public class InquiryServiceImpl implements InquiryService {
             throw new NotFoundException("수정할 게시물이 없습니다.");
         }
 
+        OAuth2User oAuth2User = (OAuth2User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String email = (String)oAuth2User.getAttributes().get("email");
+
+        if (!existing.getEmail().equals(email)) {
+            throw new ForbiddenException("접근 권한이 없습니다.");
+        }
+
         inquiryDto.setInquiry_update_date(LocalDateTime.now());
 
         inquiryMapper.updateInquiry(inquiryDto);
@@ -68,6 +84,13 @@ public class InquiryServiceImpl implements InquiryService {
 
         if (existing == null) {
             throw new NotFoundException("삭제할 게시물이 없습니다.");
+        }
+
+        OAuth2User oAuth2User = (OAuth2User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String email = (String)oAuth2User.getAttributes().get("email");
+
+        if (!existing.getEmail().equals(email)) {
+            throw new ForbiddenException("접근 권한이 없습니다.");
         }
 
         inquiryMapper.deleteInquiry(inquiry_id);
